Args {gdata}	R Documentation

Describe Function Arguments

Description

Display function argument names and corresponding default values, formatted in two columns for easy reading.

Usage

Args(name, sort=FALSE)

Arguments

name
a function or function name.

sort
whether arguments should be sorted.

Value

A data frame with named rows and a single column called value, containing the default value of each argument.

Note

Primitive functions like sum and all have no formal arguments. See the formals help page.

Author(s)

Arni Magnusson

See Also

Args is a verbose alternative to args, based on formals.

help also describes function arguments.

Examples

Args(glm)
Args(scan)
Args(legend, sort=TRUE)

[Package gdata version 2.18.0 Index]