env {gdata}	R Documentation

Describe All Loaded Environments

Description

Display name, number of objects, and size of all loaded environments.

Usage

env(unit="KB", digits=0)

Arguments

unit
unit for displaying environment size: "bytes", "KB", "MB", or first letter.

digits
number of decimals to display when rounding environment size.

Value

A data frame with the following columns:

Environment
environment name.

Objects
number of objects in environment.

KB
environment size (see notes).

Note

The name of the environment size column is the same as the unit used.

Author(s)

Arni Magnusson

See Also

env is a verbose alternative to search.

ll is a related function that describes objects in an environment.

Examples

## Not run: 
env()

## End(Not run)

[Package gdata version 2.18.0 Index]