gdata-package {gdata}	R Documentation

Various R Programming Tools for Data Manipulation

Description

Various R programming tools for data manipulation, including:

Medical unit conversions: ConvertMedUnits, MedUnits

Combining objects: link{bindData}, cbindX, combine, interleave

Character vector operations: centerText, startsWith, trim

Factor manipulation: levels, reorder.factor, mapLevels

Obtaining information about R objects: object_size, env, humanReadable, is.what, ll, keep, ls.funs, Args, nPairs, nobs

Generating fixed-width format files: write.fwf

Extracting components of date & time objects: getYear, getMonth, getDay, getHour, getMin, getSec

Operations on columns of data frames: matchcols, rename.vars

Matrix operations: unmatrix, upperTriangle, lowerTriangle

Operations on vectors: case, unknownToNA, duplicated2, trimSum

Operations on data frames: frameApply, wideByFactor

Value of last evaluated expression: ans

Wrapper for sample that ensures consistent behavior for both scalar and vector arguments: resample

Note

browseVignettes() shows package vignettes.

Author(s)

Gregory R. Warnes, Gregor Gorjanc, Arni Magnusson, Liviu Andronic, Jim Rogers, Don MacQueen, and Ales Korosec, with contributions by Ben Bolker, Michael Chirico, Gabor Grothendieck, Thomas Lumley, and Brian Ripley.

[Package gdata version 3.0.1 Index]