mv {gdata}	R Documentation

Rename an Object

Description

Rename an object.

Usage

mv(from, to, envir = parent.frame())

Arguments

from
Character scalar giving the source object name

to
Character scalar giving the desination object name

envir
Environment in which to do the rename

Details

This function renames an object by the value of object a to the name b, and removing a.

Value

Invisibly returns the value of the object.

Author(s)

Gregory R. Warnes greg@warnes.net

See Also

rm, assign

Examples

  a <- 1:10
  a
  mv("a", "b")
  b
  exists("a")

[Package gdata version 2.18.0 Index]