.runRUnitTestsGdata {gdata}	R Documentation

Run RUnit tests for the gdata package

Description

Run RUnit tests to perform the validity/unit testing of installed gdata package on your own computer.

Usage

.runRUnitTestsGdata(testFileRegexp="^runit.+\\.[rR]$")

Arguments

testFileRegexp 	 regular expression; see details

Details

Argument testFileRegexp can be used to specify different sets of tests provided by the package. The following values are sensible:

"^runit.+\\.[rR]$" for basic tests

Value

None, just the print out of RUnit testing.

See Also

defineTestSuite in RUnit package

Examples

  ## Basic testing
  .runRUnitTestsGdata()

[Package gdata version 2.6.1 Index]