sheetCount {gdata}	R Documentation

Count or list sheet names in Excel spreadsheet files.

Description

Count or list sheet names in Excel spreadsheet files.

Usage

sheetCount(xls, verbose = FALSE, perl = "perl")
sheetNames(xls, verbose = FALSE, perl = "perl")

Arguments

xls
File path to spreadsheet. Supports "http://", "https://", and "ftp://" URLS.

verbose
If TRUE, show additional messages during processing.

perl
Path to perl interpreter.

Value

sheetCount returns the number of sheets in the spreadsheet. sheetNames returns the names of sheets in the spreadsheet.

See Also

read.xls, xls2csv.

Examples

   xlsfile <- system.file("xls", "iris.xls", package = "gdata")
   xlsfile

   sheetCount(xlsfile)

   exampleFile <- file.path(path.package('gdata'),'xls',
                            'ExampleExcelFile.xls')
   exampleFile2007 <- file.path(path.package('gdata'),'xls',
                            'ExampleExcelFile.xlsx')

   sheetCount(exampleFile)

   if( 'XLSX' %in% xlsFormats() )  # if XLSX is supported..
     sheetNames(exampleFile2007)

[Package gdata version 2.18.0 Index]