Args {gdata}	R Documentation

Formatted Argument List of a Function

Description

Display function argument names and corresponding default values, formatted in two columns for easy reading.

Usage

Args(name, sort.args = FALSE)

Arguments

name 	 a function or function name.

sort.args 	 whether arguments should be sorted.

Value

A data frame with named rows and a single column called value, containing the default value of each argument.

Author(s)

Arni Magnusson arnima@u.washington.edu

See Also

args, formals, help.

Examples

Args(glm)
Args(scan)
Args(legend)

[Package gdata version 2.6.1 Index]