env {gdata}	R Documentation

Display Information about All Loaded Environments

Description

Display name, number of objects, and size of all loaded environments.

Usage

env(unit="KB", digits=0)

Arguments

unit 	 required unit for displaying environment size: "bytes", "KB", "MB", or first letter.

digits 	 number of decimals to display when rounding environment size.

Details

A verbose alternative to search().

Value

A data frame with the following columns:

Environment 	 environment name.

Objects 	 number of objects in environment.

KB 	 environment size (see notes).

Note

The name of the environment size column is the same as the unit used.

Author(s)

Arni Magnusson arnima@u.washington.edu

See Also

search displays environment names.

ll is related to env.

Examples

## Not run: 
env()
## End(Not run)

[Package gdata version 2.6.1 Index]