gdata-defunct {gdata}	R Documentation

Defunct Functions in Package 'gdata'

Description

The functions or variables listed here are no longer part of 'gdata'.

Usage

aggregate.table(x, by1, by2, FUN=mean, ...)

Arguments

x
data to be summarized

by1
first grouping factor.

by2
second grouping factor.

FUN
a scalar function to compute the summary statistics which can be applied to all data subsets. Defaults to mean.

...
Optional arguments for FUN.

Details

aggregate.table(x, by1, by2, FUN=mean, ...) should be replacede by tapply(X=x, INDEX=list(by1, by2), FUN=FUN, ...).

[Package gdata version 2.18.0 Index]