gdata-package {gdata}	R Documentation

Various R programming tools for data manipulation

Description

gdata package provides various R programming tools for data manipulation.

Details

The following are sources of information on gdata package:

DESCRIPTION file	 library(help="gdata")

This file	 package?gdata

Vignette	 vignette("unknown")

vignette("mapLevels")

Some help files	 read.xls

write.fwf

News	 file.show(system.file("NEWS", package="gdata"))

[Package gdata version 2.18.0 Index]