gdata-package {gdata}	R Documentation

Various R programming tools for data manipulation

Description

gdata package provides various R programming tools for data manipulation.

Details

The following are sources of information on gdata package:

DESCRIPTION file 	 library(help="gdata")

This file 	 package?gdata

Vignette 	 vignette("unknown")

vignette("mapLevels")

Some help files 	 read.xls

write.fwf

News 	 file.show(system.file("NEWS", package="gdata"))

Testing

If you want to perform the validity/unit testing of the installed ggmisc package on your own computer, take a look at .runRUnitTestsGdata function - please note that you need the RUnit package for this to work.

[Package gdata version 2.6.1 Index]