ll {gdata}	R Documentation

Display Information about Objects or Elements

Description

Display name, class, size, and dimensions of each object in a given environment. Alternatively, if the main argument is a list-like object, its elements are listed and described.

Usage

ll(pos=1, unit="KB", digits=0, dimensions=FALSE, function.dim="",
   sort.elements=FALSE, ...)

Arguments

pos 	 environment position number, environment name, data frame, list, model, or any object that is.list().

unit 	 required unit for displaying object size: "bytes", "KB", "MB", or first letter.

digits 	 number of decimals to display when rounding object size.

dimensions 	 whether object dimensions should be returned.

function.dim 	 value to report as the dimension of function objects.

sort.elements 	 whether elements should be sorted by name.

... 	 passed to ls().

Details

A verbose alternative to ls() and names().

Value

A data frame with named rows and the following columns:

Class 	 object class.

KB 	 object size (see notes).

Dim 	 object dimensions (optional).

Note

The name of the object size column is the same as the unit used.

Author(s)

Arni Magnusson arnima@u.washington.edu, with a contribution by Jim Rogers james.a.rogers@pfizer.com

See Also

ls displays names of objects in a given environment.

names, str, and summary display different information about list-like elements.

env is related to ll.

Examples

ll()
ll(all=TRUE)
ll("package:base")
ll("package:base")[ll("package:base")$Class!="function",]

data(infert)
ll(infert)
model <- glm(case~spontaneous+induced, family=binomial, data=infert)
ll(model, dim=TRUE)
ll(model, sort=TRUE)
ll(model$family)

[Package gdata version 2.6.1 Index]