object.size {gdata}	R Documentation

Report the Space Allocated for an Object

Description

Provides an estimate of the memory that is being used to store an R object.

Usage

object.size(...)

## S3 method for class 'object_size':
print(x, quote=FALSE, humanReadable, ...)

Arguments

... 	 object.size: R objects; print; arguments to be passed to or from other methods.

x 	 output from object.size

quote 	 logical, indicating whether or not the result should be printed with surrounding quotes.

humanReadable 	 logical, use the “human readable” format.

Details

This is a modified copy from the utils package in R as fo 2008-12-15.

Exactly which parts of the memory allocation should be attributed to which object is not clear-cut. This function merely provides a rough indication: it should be reasonably accurate for atomic vectors, but does not detect if elements of a list are shared, for example. (Sharing amongst elements of a character vector is taken into account, but not that between character vectors in a single object.)

The calculation is of the size of the object, and excludes the space needed to store its name in the symbol table.

Associated space (e.g. the environment of a function and what the pointer in a EXTPTRSXP points to) is not included in the calculation.

Object sizes are larger on 64-bit platforms than 32-bit ones, but will very likely be the same on different platforms with the same word length and pointer size.

Class of returned object is c("byte", "numeric") with appropriate print and c methods.

By default object.size outputs size in bytes, but human readable format similar to ls, df or du shell commands can be invoked with options(humanReadable=TRUE).

Value

An object of class "object.size" with a length-one double value, an estimate of the memory allocation attributable to the object in bytes.

See Also

Memory-limits for the design limitations on object size. humanReadable for human readable format.

Examples

object.size(letters)
object.size(ls)
## find the 10 largest objects in the base package
z <- sapply(ls("package:base"), function(x)
            object.size(get(x, envir = baseenv())))
(tmp <- as.matrix(rev(sort(z))[1:10]))

as.object_size(14567567)
options(humanReadable=TRUE)
(z <- object.size(letters, c(letters, letters), rep(letters, 100), rep(letters, 10000)))
is.object_size(z)
as.object_size(14567567)

[Package gdata version 2.6.1 Index]