read.xls {gdata}	R Documentation

Read Excel files

Description

Reads a Microsoft Excel file into a data frame

Usage


read.xls(xls, sheet=1, verbose=FALSE, ..., perl="perl")

Arguments

xls 	 name of the Microsoft Excel file

sheet 	 number of sheet within the Excel file from which data are to be read

verbose 	 logical flag indicating whether details should be printed as the file is processed.

perl 	 name of the perl executable to be called.

... 	 additional arguments to read.table. The defaults of read.csv are used.

Details

This function works translating the named Microsoft Excel file into a temporary .csv file, using Greg Warnes' xls2csv Perl script (installed as part of the gregmisc package).

Caution: In the conversion to csv, strings will be quoted. This can be problem if you are trying to use the comment.char option of read.table since the first character of all lines (including comment lines) will be """ after conversion.

Value

a data frame

Note

Either a working version of Perl must be present in the executable search path, or the exact path of the perl executable must be provided via the perl argument. See the examples below for an illustration.

Author(s)

Jim Rogers james.a.rogers@pfizer.com, modified and extended by Gregory R. Warnes warnes@bst.rochester.edu.

References

http://www.analytics.washington.edu/statcomp/downloads/xls2csv

See Also

read.csv

Examples



   # iris.xls is included in the gregmisc package for use as an example
   xlsfile <- file.path(.path.package('gdata'),'xls','iris.xls')
   xlsfile

   iris <- read.xls(xlsfile)
   head(iris)  # look at the top few rows

  ## Not run: 
   # Example specifying exact Perl path for default MS-Windows install of
   # ActiveState perl
   iris <- read.xls(xlsfile, perl="C:\perl\bin\perl.exe")

   # Example specifying exact Perl path for Unix systems
   iris <- read.xls(xlsfile, perl="/usr/bin/perl")
   ## End(Not run)

[Package gdata version 2.3.1 Index]