resample {gdata}	R Documentation

Consistent Random Samples and Permutations

Description

resample takes a sample of the specified size from the elements of x using either with or without replacement.

Usage

resample(x, size, replace = FALSE, prob = NULL)

Arguments

x 	 A numeric, complex, character or logical vector from which to choose.

size 	 Non-negative integer giving the number of items to choose.

replace 	 Should sampling be with replacement?

prob 	 A vector of probability weights for obtaining the elements of the vector being sampled.

Details

resample differs from the S/R sample function in resample always considers x to be a vector of elements to select from, while sample treats a vector of length one as a special case and samples from 1:x. Otherwise, the functions have identical behavior.

Value

Vector of the same length as the input, with the elements permuted.

Author(s)

Gregory R. Warnes greg@random-technologies-llc.com

See Also

sample

Examples

## sample behavior differs if first argument is scalar vs vector
sample( c(10) )
sample( c(10,10) )

## resample has the consistent behavior for both cases
resample( c(10) )
resample( c(10,10) )

[Package gdata version 2.6.1 Index]